requirements.txt
image==1.5.5
numpy==1.12.1
PyOpenGL==3.1.1
PyOpenGL-accelerate==3.1.1
